requirements.txt
google.generativeai
streamlit
python-dotenv
pyPDF
faiss-cpu
langchain
langchain-huggingface
langchain_community
sentence-transformers